veterinary-appointment-system/auth-service/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
bcrypt==4.0.1
python-dotenv==1.0.0
marshmallow==3.20.1